requirements.txt
pytest==8.1.1
python-dotenv==1.0.1
allure-pytest==2.13.5
requests==2.31.0
jsonschema==4.22.0